some mistakes left over from merge

+++ interfaces/src/main/java/ar/edu/itba/paw/interfaces/dao/PropertyDao.java
@@ -5,6 +5,6 @@ import ar.edu.itba.paw.model.User;
 
 public interface PropertyDao extends Dao<Property>{
 	Property get(long id);
-	boolean showInterest(int propertyId, User user);
+	boolean showInterest(long propertyId, User user);
 	Property getPropertyWithRelatedEntities(long id);
 }
