+++ service/src/main/java/ar/edu/itba/paw/service/APPropertyService.java
@@ -36,7 +36,7 @@ public class APPropertyService implements PropertyService {
     }
 
     @Override
-    public List<String> showInterestOrReturnErrors(int propertyId, String username) {
+    public List<String> showInterestOrReturnErrors(long propertyId, String username) {
         List<String> errors = new LinkedList<>();
         User user = userDao.getByUsername(username);
         CheckUserAndPropertyExist(propertyId, errors, user);
@@ -48,7 +48,7 @@ public class APPropertyService implements PropertyService {
         return errors;
     }
 
-    private void CheckUserAndPropertyExist(int propertyId, List<String> errors, User user) {
+    private void CheckUserAndPropertyExist(long propertyId, List<String> errors, User user) {
         if (propertyDao.get(propertyId) == null)
             errors.add(PROPERTY_NOT_FOUND);
         if(user == null)
